perf(moe): build the tokens mask in the routing weights' dtype
`get_tokens_mask` passed Python floats to `torch.where`, so the mask came
back in the torch default dtype (fp32). Multiplying it into the bf16
routing weights promoted the whole [E, T] tensor, which the compiler reads
as a precision island and runs in dlfp16 -- three device casts per layer to
carry a mask of ones and zeros.

Take a `dtype` argument and materialize the branch values in it, and pass
`masked_routing_weights.dtype` from both call sites. The mask follows the
caller rather than being pinned to bf16, so a model whose routing weights
are fp32 still gets an fp32 mask and stays cast-free. `dtype=None` keeps
the previous behavior.

Bit-identical: 1.0/0.0 (and -inf) are exact in bf16, and multiplying by
them is exact in any float format.

+++ vllm_rbln/model_executor/layers/fused_moe/runner/moe_runner.py
@@ -262,7 +262,9 @@ class RBLNMoERunner(MoERunner):
             use_moe_tokens_mask = envs.VLLM_RBLN_USE_MOE_TOKENS_MASK
             if use_moe_tokens_mask:
                 tokens_mask = get_tokens_mask(
-                    max_pad, device=masked_routing_weights.device
+                    max_pad,
+                    device=masked_routing_weights.device,
+                    dtype=masked_routing_weights.dtype,
                 ).transpose(1, 0)  # [1, R*max_pad]
                 masked_routing_weights = masked_routing_weights * tokens_mask
 
@@ -480,7 +482,9 @@ class RBLNMoERunner(MoERunner):
         use_moe_tokens_mask = envs.VLLM_RBLN_USE_MOE_TOKENS_MASK
         if use_moe_tokens_mask:
             tokens_mask = get_tokens_mask(
-                num_tokens, device=masked_routing_weights.device
+                num_tokens,
+                device=masked_routing_weights.device,
+                dtype=masked_routing_weights.dtype,
             ).transpose(1, 0)  # [1, t]
 
             # [t, E] * [t, 1] (broadcast)
